src/main/java/core/parallax/ScreenHook.java
package core.parallax;

// (0, 0) is at the center of the screen
public class ScreenHook {
    private int centerX;
    private int centerY;
    private int width;
    private int height;

    public ScreenHook(int centerX, int centerY, int width, int height) {
        this.centerX = centerX;
        this.centerY = centerY;
        this.width = width;
        this.height = height;
    }

    public int centerX() {
        return centerX;
    }

    public int centerY() {
        return centerY;
    }

    public int width() {
        return width;
    }

    public int height() {
        return height;
    }

    public void setCenterX(int centerX) {
        this.centerX = centerX;
    }

    public void setCenterY(int centerY) {
        this.centerY = centerY;
    }

    public void setWidth(int width) {
        this.width = width;
    }

    public void setHeight(int height) {
        this.height = height;
    }
}
